refactor: Generaize approve component
